修复菜单滑动条控件数值在文本模式下显示不正常的问题 修复了不同地方调用菜单跳转导致结果偏移错位的问题

+++ Menu.h
@@ -60,10 +60,10 @@ extern struct Smooth_Animation Menu_Smooth_Animation[];
 		step 步进
 */
 struct Slide_Bar {
-	int* x;	 //值
-	int min;
-	int max;
-	int step;
+	float* x;	 //值
+	float min;
+	float max;
+	float step;
 };
 extern struct Slide_Bar Slide_space[];
